+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 
 ### Added
 - Rewards and sovereign rewards are now stored in memory.
+- Actions to serve both rewards as JSON.
+
+### Removed
+- Saving rewards as file.